Day 22B Implementation via Reddit

+++ src/main/java/org/jgoeres/adventofcode2019/Day22/AbstractShuffle.java
@@ -1,7 +1,7 @@
 package org.jgoeres.adventofcode2019.Day22;
 
 abstract class AbstractShuffle implements Shuffle {
-    protected int deckSize;
+    protected long deckSize;
 
     public AbstractShuffle() {
     }
